Rename danish translation file to match the others
