+++ langs/dk.h
